Money/requirements.txt
# Core AI
crewai[google-genai]>=0.110.0
crewai-tools>=0.38.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langsmith>=0.2.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9

# Integrations
twilio>=9.3.0
composio-crewai>=0.6.0
requests>=2.32.0
aiohttp>=3.10.0

# Reliability
itsdangerous>=2.1.0
tenacity>=9.0.0
python-dotenv>=1.0.0
rich>=13.7.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
prometheus-client>=0.20.0
